UI to add new group and improve structure of UI
